fix a bunch of warnings (reported by cparser)

+++ attribute.c
@@ -227,7 +227,7 @@ static void handle_attribute_aligned(const attribute_t *attribute,
 		break;
 	case ENTITY_STRUCT:
 	case ENTITY_UNION:
-		if (alignment > entity->compound.alignment) {
+		if (alignment > (int)entity->compound.alignment) {
 			entity->compound.alignment = alignment;
 		}
 		break;
